+++ README.md
@@ -26,6 +26,61 @@
 <br>
 <br>
 
+2022.12 ver.
+
+## useLocalStorage 훅 만들어서 로컬스토리지를 통해 투두리스트 데이터 관리
+
+App에서 useLocalStorage 훅을 불러와서, 하위 컴포넌트들에게 전달하여 state 값 변경함
+
+```js
+// hooks/useLocalStorage.js
+import { useState, useEffect } from "react";
+
+const useLocalStorage = (key, initialValue) => {
+  const localDB = JSON.parse(localStorage.getItem(key));
+
+  const [state, setState] = useState(localDB || initialValue);
+
+  useEffect(() => {
+    localStorage.setItem(key, JSON.stringify(state));
+  }, [state]);
+  return [state, setState];
+};
+export default useLocalStorage;
+```
+
+```js
+// App.js
+
+function App() {
+  const [todos, setTodos] = useLocalStorage("todoApp", initialState);
+
+  const [createMode, setCreate] = useState < boolean > false;
+
+  return (
+    <div className="App">
+      <div className="container">
+        <div className="inner">
+          <Header todos={todos} />
+          <TodoList todos={todos} setTodos={setTodos} />
+          <TodoInsert
+            createMode={createMode}
+            setCreate={setCreate}
+            todos={todos}
+            setTodos={setTodos}
+          />
+          <CreateBtn createMode={createMode} setCreate={setCreate} />
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default App;
+```
+
+2022.11 ver.
+
 ## todo list 로컬스토리지에서 관리
 
 1. 첫 로드 시 로컬스토리지에서 해당 투두 리스트 가져오기 (비동기)
