+++ src/components/TodoList.tsx
@@ -1,14 +1,14 @@
 import React, { Dispatch, SetStateAction } from "react";
 import TodoItem from "./TodoItem";
 
-import { TodosState } from "../modules/type";
+import { TodosProps } from "../type";
 
 function TodoList({
   todos,
   setTodos,
 }: {
-  todos: TodosState;
-  setTodos: Dispatch<SetStateAction<TodosState>>;
+  todos: TodosProps;
+  setTodos: Dispatch<SetStateAction<TodosProps>>;
 }) {
   if (todos.length === 0) return <p className="empty-todo">할일이 없습니다.</p>;
   return (
@@ -18,7 +18,12 @@ function TodoList({
           {todos
             .sort((a, b) => (a.done > b.done ? 1 : -1))
             .map((todo) => (
-              <TodoItem todo={todo} key={todo.id} />
+              <TodoItem
+                todo={todo}
+                key={todo.id}
+                setTodos={setTodos}
+                todos={todos}
+              />
             ))}
         </ul>
       )}
