+++ src/hooks/useLocalStorage.js
@@ -2,10 +2,11 @@ import { useState, useEffect } from "react";
 
 const useLocalStorage = (key, initialValue) => {
   const localDB = JSON.parse(localStorage.getItem(key));
+
   const [state, setState] = useState(localDB || initialValue);
-  console.log(key, localDB, initialValue);
+
   useEffect(() => {
-    localStorage.setItem(key, JSON.stringify(initialValue));
+    localStorage.setItem(key, JSON.stringify(state));
   }, [state]);
   return [state, setState];
 };
